refactor: remove unnecessary types , enum and function
