+++ packages/ui/src/lib/components/organisms/rating-history/rating-history.stories.svelte
@@ -1,5 +1,5 @@
 <script module lang="ts">
-	import { Status } from '$lib/utils.js';
+	import type { ReviewStatus } from '@cugetreg/zod-schemas/constants';
 
 	import { defineMeta } from '@storybook/addon-svelte-csf';
 
@@ -45,7 +45,7 @@
 				code: '0123104',
 				name: 'CON PDG PEACE CONFWV',
 				tag: 'หมวดมนุษย์',
-				status: Status.APPROVED,
+				status: 'APPROVED',
 				rating: 4,
 				term: 'ภาคต้น 2566'
 			},
@@ -53,7 +53,7 @@
 				code: '0123101',
 				name: 'PARAGRAPH WRITING',
 				tag: 'หมวดมนุษย์',
-				status: Status.APPROVED,
+				status: 'APPROVED',
 				rating: 3,
 				term: 'ภาคปลาย 2565'
 			},
@@ -61,7 +61,7 @@
 				code: '0123101',
 				name: 'PARAGRAPH WRITING',
 				tag: 'หมวดมนุษย์',
-				status: Status.APPROVED,
+				status: 'APPROVED',
 				rating: 3,
 				term: 'ภาคปลาย 2566'
 			},
@@ -69,7 +69,7 @@
 				code: '2190201',
 				name: 'COM PROG',
 				tag: 'หมวดวิทย์',
-				status: Status.PENDING,
+				status: 'PENDING',
 				rating: 2.5,
 				term: 'ภาคปลาย 2566'
 			}
