+++ packages/ui/src/lib/utils.ts
@@ -11,8 +11,3 @@ export type WithoutChildren<T> = T extends { children?: any } ? Omit<T, 'childre
 export type WithoutChildrenOrChild<T> = WithoutChildren<WithoutChild<T>>;
 export type WithElementRef<T, U extends HTMLElement = HTMLElement> = T & { ref?: U | null };
 
-export enum Status {
-	PENDING = 'PENDING',
-	APPROVED = 'APPROVED',
-	REJECTED = 'REJECTED'
-}
